Add widget boolean/checkbox control
