+++ src/ui/widget/index.ts
@@ -1,2 +1,3 @@
+export * from './types';
 export { WidgetManager } from './widget-manager';
 export { WidgetView } from './widget-view';
