+++ src/ui/widget/types.ts
@@ -1,0 +1,3 @@
+export interface WidgetParamControl {
+    
+}
